Add files via upload
This code generates a sample of 3000 observations from a mixture of two gamma distributions and one normal distribution using the numpy and scipy libraries. It then creates a histogram of the sample using matplotlib and fits an empirical cumulative distribution function (ECDF) to the sample using the statsmodels library. Finally, the code calculates the probabilities of various values using the ECDF and prints them to the console.